# 02_docs/microservices/secondhand-service/openapi.yaml
openapi: 3.0.3
info:
  title: Kinda Goods secondhand-service API
  version: 1.0.0
  description: >-
    MS-04 contract for UC16-UC19. UC20 order, payment and logistics APIs remain
    owned by order-service. All protected APIs derive the caller from JWT and
    ignore identity supplied by clients.
servers:
  - url: /
tags:
  - name: Product
  - name: Bargain
  - name: Auction
  - name: Internal events
paths:
  /api/secondhand/list:
    get:
      tags: [Product]
      summary: List publicly tradable secondhand products
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - { name: keyword, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: integer } }
        - { name: minPrice, in: query, schema: { type: number } }
        - { name: maxPrice, in: query, schema: { type: number } }
        - { name: conditionLevel, in: query, schema: { type: string } }
        - { name: isNegotiable, in: query, schema: { type: integer, enum: [0, 1] } }
        - { name: sortBy, in: query, schema: { type: string, enum: [latest, priceAsc, priceDesc] } }
      responses:
        '200': { $ref: '#/components/responses/ProductPage' }
  /api/secondhand/detail/{productId}:
    get:
      tags: [Product]
      summary: Get a public product detail
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200': { $ref: '#/components/responses/Product' }
        '404': { $ref: '#/components/responses/Error' }
  /api/secondhand/seller-public/{sellerUserId}:
    get:
      tags: [Product]
      summary: Get the public secondhand seller profile
      parameters:
        - name: sellerUserId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Public seller profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/SellerPublic' }
  /api/secondhand/seller-public/{sellerUserId}/products:
    get:
      tags: [Product]
      summary: List one seller's publicly tradable products
      parameters:
        - name: sellerUserId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/ProductPage' }
  /api/secondhand/seller/list:
    get:
      tags: [Product]
      summary: List products owned by the authenticated seller
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer, enum: [1, 2, 3, 4] } }
      responses:
        '200': { $ref: '#/components/responses/ProductPage' }
        '401': { $ref: '#/components/responses/Error' }
  /api/secondhand/seller:
    post:
      tags: [Product]
      summary: Publish a secondhand product in risk-pending state
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductSave' }
      responses:
        '201': { $ref: '#/components/responses/Product' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/secondhand/seller/{productId}:
    put:
      tags: [Product]
      summary: Edit an owned product
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProductId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductSave' }
      responses:
        '200': { $ref: '#/components/responses/Product' }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Product]
      summary: Delete an owned product that has no active trade
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProductId' }]
      responses:
        '200': { $ref: '#/components/responses/Empty' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/seller/{productId}/status:
    post:
      tags: [Product]
      summary: Put an owned product on or off shelf
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProductId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: integer, enum: [1, 2] }
      responses:
        '200': { $ref: '#/components/responses/Product' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/{productId}/buy:
    post:
      tags: [Product]
      summary: Atomically reserve a product and request a pending-payment order
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProductId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [addressId]
              properties:
                addressId: { type: integer, format: int64 }
                remark: { type: string, maxLength: 255 }
      responses:
        '200': { $ref: '#/components/responses/TradeOrder' }
        '202': { $ref: '#/components/responses/TradeOrder' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/bargain/apply:
    post:
      tags: [Bargain]
      summary: Submit an idempotent bargain request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, sellerUserId, proposedPrice]
              properties:
                productId: { type: integer, format: int64 }
                sellerUserId: { type: integer, format: int64 }
                proposedPrice: { type: number, minimum: 0.01 }
      responses:
        '200': { $ref: '#/components/responses/Negotiation' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/bargain/confirm:
    post:
      tags: [Bargain]
      summary: Seller accepts through CAS and creates at most one order
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [negotiationId, confirmedPrice]
              properties:
                negotiationId: { type: integer, format: int64 }
                confirmedPrice: { type: number, minimum: 0.01 }
                createOrder: { type: boolean, default: true }
      responses:
        '200': { $ref: '#/components/responses/Negotiation' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/bargain/{negotiationId}/reject:
    post:
      tags: [Bargain]
      summary: Seller rejects a pending bargain
      security: [{ bearerAuth: [] }]
      parameters:
        - name: negotiationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/Negotiation' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/bargain/list:
    get:
      tags: [Bargain]
      summary: List bargains involving the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - { name: productId, in: query, schema: { type: integer, format: int64 } }
        - { name: counterpartUserId, in: query, schema: { type: integer, format: int64 } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/NegotiationPage' }
  /api/secondhand/trade/bargain/effective:
    get:
      tags: [Bargain]
      summary: Get the buyer's effective bargain for a product
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/Negotiation' }
  /api/secondhand/trade/auction:
    post:
      tags: [Auction]
      summary: Create an auction for an owned product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, startPrice, incrementAmount, durationMinutes]
              properties:
                productId: { type: integer, format: int64 }
                startPrice: { type: number, minimum: 0.01 }
                incrementAmount: { type: number, minimum: 0.01 }
                durationMinutes: { type: integer, minimum: 1, maximum: 10080 }
      responses:
        '200': { $ref: '#/components/responses/Auction' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/auction/product/{productId}:
    get:
      tags: [Auction]
      summary: Get the auction and bid history for a product
      parameters: [{ $ref: '#/components/parameters/ProductId' }]
      responses:
        '200': { $ref: '#/components/responses/Auction' }
  /api/secondhand/trade/auction/seller/list:
    get:
      tags: [Auction]
      summary: List auctions owned by the authenticated seller
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/AuctionPage' }
  /api/secondhand/trade/auction/{auctionId}/bid:
    post:
      tags: [Auction]
      summary: Place a concurrency-safe bid
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/AuctionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bidAmount]
              properties:
                bidAmount: { type: number, minimum: 0.01 }
      responses:
        '200': { $ref: '#/components/responses/Auction' }
        '409': { $ref: '#/components/responses/Error' }
  /api/secondhand/trade/auction/{auctionId}/close:
    post:
      tags: [Auction]
      summary: Seller closes and idempotently settles an auction
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/AuctionId' }]
      responses:
        '200': { $ref: '#/components/responses/Auction' }
  /api/secondhand/trade/auction/{auctionId}/flow:
    post:
      tags: [Auction]
      summary: Mark an auction without bids as unsold
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/AuctionId' }]
      responses:
        '200': { $ref: '#/components/responses/Auction' }
  /internal/events/order-status-changed:
    post:
      tags: [Internal events]
      summary: Idempotently consume OrderStatusChanged.v1
      security: [{ internalToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, orderBusinessKey, orderId, newStatus]
              properties:
                eventId: { type: string }
                orderBusinessKey: { type: string }
                orderId: { type: integer, format: int64 }
                newStatus: { type: string }
      responses:
        '200': { $ref: '#/components/responses/EventConsumed' }
        '401': { $ref: '#/components/responses/Error' }
  /internal/events/product-risk-decided:
    post:
      tags: [Internal events]
      summary: Idempotently consume ProductRiskDecided.v1
      security: [{ internalToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, productId, decision]
              properties:
                eventId: { type: string }
                productId: { type: integer, format: int64 }
                decision: { type: string, enum: [APPROVED, REJECTED] }
      responses:
        '200': { $ref: '#/components/responses/EventConsumed' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    AuctionId:
      name: auctionId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    PageNum:
      name: pageNum
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
  responses:
    Product:
      description: Product envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Product' }
    ProductPage:
      description: Product page envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Page' }
    Negotiation:
      description: Negotiation envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Negotiation' }
    NegotiationPage:
      description: Negotiation page envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Page' }
    Auction:
      description: Auction envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Auction' }
    AuctionPage:
      description: Auction page envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Page' }
    TradeOrder:
      description: Existing, created or retrying order request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/TradeOrder' }
    EventConsumed:
      description: Event consumption result
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      status: { type: string, enum: [CONSUMED, DUPLICATE] }
    Empty:
      description: Successful response without data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error:
      description: Business or authentication error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { nullable: true }
    Page:
      type: object
      properties:
        records: { type: array, items: { type: object } }
        total: { type: integer, format: int64 }
        pageNum: { type: integer, format: int64 }
        pageSize: { type: integer, format: int64 }
    ProductSave:
      type: object
      required: [name, images, originPrice, salePrice, categoryId, subCategoryId, conditionLevel, isNegotiable]
      properties:
        name: { type: string, maxLength: 120 }
        cover: { type: string, maxLength: 255 }
        images: { type: array, minItems: 1, maxItems: 9, items: { type: string } }
        description: { type: string, maxLength: 2000 }
        originPrice: { type: number, minimum: 0.01 }
        salePrice: { type: number, minimum: 0.01 }
        categoryId: { type: integer }
        subCategoryId: { type: integer }
        conditionLevel: { type: string, maxLength: 30 }
        isNegotiable: { type: integer, enum: [0, 1] }
        status: { type: integer, enum: [1, 2] }
    Product:
      type: object
      properties:
        id: { type: integer, format: int64 }
        sellerUserId: { type: integer, format: int64 }
        sellerName: { type: string }
        name: { type: string }
        salePrice: { type: number }
        status: { type: integer, enum: [1, 2, 3, 4] }
        conditionLevel: { type: string }
        isNegotiable: { type: integer, enum: [0, 1] }
        risk:
          type: object
          properties:
            auditStatus: { type: string }
            suggestion: { type: string }
    SellerPublic:
      type: object
      properties:
        userId: { type: integer, format: int64 }
        nickname: { type: string }
        avatar: { type: string }
        region: { type: string }
        rating:
          type: object
          properties:
            score: { type: integer }
            level: { type: string }
            completedTrades: { type: integer, format: int64 }
    Negotiation:
      type: object
      properties:
        id: { type: integer, format: int64 }
        productId: { type: integer, format: int64 }
        buyerUserId: { type: integer, format: int64 }
        sellerUserId: { type: integer, format: int64 }
        proposedPrice: { type: number }
        confirmedPrice: { type: number, nullable: true }
        status: { type: string, enum: [PENDING, ACCEPTING, ACCEPTED, REJECTED, FAILED] }
        usedOrderId: { type: integer, format: int64, nullable: true }
    Auction:
      type: object
      properties:
        id: { type: integer, format: int64 }
        productId: { type: integer, format: int64 }
        sellerUserId: { type: integer, format: int64 }
        startPrice: { type: number }
        currentPrice: { type: number }
        incrementAmount: { type: number }
        highestBidderId: { type: integer, format: int64, nullable: true }
        status: { type: string, enum: [OPEN, SETTLING, FINISHED, FLOW, FAILED] }
        settledOrderId: { type: integer, format: int64, nullable: true }
        bids: { type: array, items: { type: object } }
    TradeOrder:
      type: object
      properties:
        tradeType: { type: string, enum: [DIRECT, BARGAIN, AUCTION] }
        tradeId: { type: string }
        orderBusinessKey: { type: string }
        requestStatus: { type: string, enum: [PENDING, RETRY, CREATED, FAILED] }
        orderId: { type: integer, format: int64, nullable: true }
        orderStatus: { type: string, example: PENDING_PAY }